requirements.txt
pandas
numpy
pyarrow
fastparquet
requests
pyyaml
python-dotenv
google-generativeai
scikit-learn
matplotlib
seaborn
jupyter
